requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
Django
django-filter==2.4.0
djangorestframework
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
Pillow==8.1.0
pycodestyle==2.4.0
PyJWT==2.0.1
pylint==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
django-cors-headers
django-allauth
django-rest-auth